fix: test flow types [#1446]
