+++ src/packages/@common/types/workspace/testflow.ts
@@ -1,3 +1,9 @@
+import type { RequestDataTypeEnum } from "./request";
+import type { Tab } from "./tab";
+
+/**
+ * TF (Test FLow) Default Type
+ */
 export interface TFNodeType {
   id: string;
   type: string;
@@ -26,43 +32,9 @@ export interface TFDocumentType {
   updatedBy: string;
 }
 
-export interface TFTabNodesWrapperType {
-  nodes: {
-    id: string;
-    type: string;
-    data: {
-      requestId: string;
-      folderId: string;
-      collectionId: string;
-      name: string;
-      method: string;
-    };
-    position: {
-      x: number;
-      y: number;
-    };
-  }[];
-}
-
-export interface TFTabEdgesWrapperType {
-  edges: {
-    id: string;
-    source: string;
-    target: string;
-  }[];
-}
-
-export interface TFTabItemType
-  extends TFTabNodesWrapperType,
-    TFTabEdgesWrapperType {}
-
 export enum TFDefaultEnum {
   FULL_NAME = "Test Flow",
   NAME = "Flow",
-}
-
-export interface TFTabItemWrapperType {
-  testflow: TFTabItemType;
 }
 
 export interface TFAPIResponseType {
@@ -113,3 +85,86 @@ export interface TFResponseStateType {
   responseBodyLanguage: string;
   responseBodyFormatter: string;
 }
+
+/**
+ * TF (Test FLow) Tab TYPE
+ */
+export interface TFTabNodesWrapperType {
+  nodes: {
+    id: string;
+    type: string;
+    data: {
+      requestId: string;
+      folderId: string;
+      collectionId: string;
+      name: string;
+      method: string;
+    };
+    position: {
+      x: number;
+      y: number;
+    };
+  }[];
+}
+
+export interface TFTabEdgesWrapperType {
+  edges: {
+    id: string;
+    source: string;
+    target: string;
+  }[];
+}
+
+export interface TFTabItemType
+  extends TFTabNodesWrapperType,
+    TFTabEdgesWrapperType {}
+
+export interface TFTabItemWrapperType {
+  testflow: TFTabItemType;
+}
+
+/**
+ * TF (Test FLow) SVELTE STORE TYPE
+ */
+export type TFKeyValueStoreType = {
+  key: string;
+  value: string;
+};
+
+export type TFHistoryAPIResponseStoreType = {
+  headers: TFKeyValueStoreType[];
+  status: string;
+  body: string;
+  time: number;
+  size: number;
+  responseContentType?: RequestDataTypeEnum;
+};
+export type TFNodeStoreType = {
+  id: string;
+  response: TFHistoryAPIResponseStoreType;
+  request: Tab;
+};
+
+export type TFAPIStoreType = {
+  method: string;
+  name: string;
+  status: string;
+  time: string;
+};
+
+export type TFHistoryStoreType = {
+  status: "pass" | "fail";
+  successRequests: string;
+  failedRequests: string;
+  totalTime: string;
+  timestamp: string;
+  requests: TFAPIStoreType[];
+  expand: boolean;
+};
+
+export type TFDataStoreType = {
+  nodes: TFNodeStoreType[];
+  history: TFHistoryStoreType[];
+  isRunHistoryEnable: boolean;
+  isTestFlowRunning: boolean;
+};
