+++ src/packages/@workspaces/features/testflow-explorer/store/testflow.ts
@@ -1,0 +1,6 @@
+import type { TFDataStoreType } from "@common/types/workspace/testflow";
+import { writable } from "svelte/store";
+
+export const testFlowDataStore = writable<Map<string, TFDataStoreType>>(
+  new Map(),
+);
